# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/upload": {
            "post": {
                "description": "上传文件并解析元数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "files"
                ],
                "parameters": [
                    {
                        "description": "上传请求信息",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ViewStruct.UploadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回metadataId",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/xml": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name to say hello",
                        "name": "filepath",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "ViewStruct.UploadRequest": {
            "type": "object",
            "properties": {
                "filePath": {
                    "description": "文件路径",
                    "type": "string"
                }
            }
        }
    }
}